backend/requirements.txt
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.2

# HTTP and JSON handling
requests==2.31.0
urllib3==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Date/time handling
python-dateutil==2.8.2

# Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
moto==4.2.14
